+++ Pulp/Common/Logging/FilterPolicies/LoggerVerbosityFilterPolicy.cpp
@@ -47,7 +47,7 @@ void LoggerVerbosityFilterPolicy::RegisterVars(void)
 }
 
 bool LoggerVerbosityFilterPolicy::Filter(const char* type, const SourceInfo& sourceInfo,
-	const char* channel, size_t verbosity, const char* format, va_list args)
+	const char* channel, int verbosity, const char* format, va_list args)
 {
 	X_UNUSED(type);
 	X_UNUSED(sourceInfo);
