+++ Pulp/Common/Logging/LoggerBase.h
@@ -19,7 +19,7 @@ public:
 	/// Destructor.
 	virtual ~LoggerBase(void);
 
-	virtual void Log(const SourceInfo& sourceInfo, const char* channel, size_t verbosity, const char* foramt, va_list args);
+	virtual void Log(const SourceInfo& sourceInfo, const char* channel, int verbosity, const char* foramt, va_list args);
 	virtual void Warning(const SourceInfo& sourceInfo, const char* channel, const char* foramt, va_list args);
 	virtual void Error(const SourceInfo& sourceInfo, const char* channel, const char* foramt, va_list args);
 	virtual void Fatal(const SourceInfo& sourceInfo, const char* channel, const char* foramt, va_list args);
@@ -38,7 +38,7 @@ public:
 
 private:
 	// derived implmentation.
-	virtual void DoLog(const SourceInfo& sourceInfo, const char* channel, size_t verbosity, const char* foramt, va_list args) X_ABSTRACT;
+	virtual void DoLog(const SourceInfo& sourceInfo, const char* channel, int verbosity, const char* foramt, va_list args) X_ABSTRACT;
 	virtual void DoWarning(const SourceInfo& sourceInfo, const char* channel, const char* foramt, va_list args) X_ABSTRACT;
 	virtual void DoError(const SourceInfo& sourceInfo, const char* channel, const char* foramt, va_list args) X_ABSTRACT;
 	virtual void DoFatal(const SourceInfo& sourceInfo, const char* channel, const char* foramt, va_list args) X_ABSTRACT;
